+++ src/controllers/commands/CommandController.cpp
@@ -949,78 +949,7 @@ void CommandController::initialize(Settings &, Paths &paths)
 
     this->registerCommand("/clip", &commands::clip);
 
-    this->registerCommand("/marker", [](const QStringList &words,
-                                        auto channel) {
-        auto *twitchChannel = dynamic_cast<TwitchChannel *>(channel.get());
-        if (twitchChannel == nullptr)
-        {
-            channel->addMessage(makeSystemMessage(
-                "The /marker command only works in Twitch channels"));
-            return "";
-        }
-
-        // Avoid Helix calls without Client ID and/or OAuth Token
-        if (getApp()->accounts->twitch.getCurrent()->isAnon())
-        {
-            channel->addMessage(makeSystemMessage(
-                "You need to be logged in to create stream markers!"));
-            return "";
-        }
-
-        // Exact same message as in webchat
-        if (!twitchChannel->isLive())
-        {
-            channel->addMessage(makeSystemMessage(
-                "You can only add stream markers during live streams. Try "
-                "again when the channel is live streaming."));
-            return "";
-        }
-
-        auto arguments = words;
-        arguments.removeFirst();
-
-        getHelix()->createStreamMarker(
-            // Limit for description is 140 characters, webchat just crops description
-            // if it's >140 characters, so we're doing the same thing
-            twitchChannel->roomId(), arguments.join(" ").left(140),
-            [channel, arguments](const HelixStreamMarker &streamMarker) {
-                channel->addMessage(makeSystemMessage(
-                    QString("Successfully added a stream marker at %1%2")
-                        .arg(formatTime(streamMarker.positionSeconds))
-                        .arg(streamMarker.description.isEmpty()
-                                 ? ""
-                                 : QString(": \"%1\"")
-                                       .arg(streamMarker.description))));
-            },
-            [channel](auto error) {
-                QString errorMessage("Failed to create stream marker - ");
-
-                switch (error)
-                {
-                    case HelixStreamMarkerError::UserNotAuthorized: {
-                        errorMessage +=
-                            "you don't have permission to perform that action.";
-                    }
-                    break;
-
-                    case HelixStreamMarkerError::UserNotAuthenticated: {
-                        errorMessage += "you need to re-authenticate.";
-                    }
-                    break;
-
-                    // This would most likely happen if the service is down, or if the JSON payload returned has changed format
-                    case HelixStreamMarkerError::Unknown:
-                    default: {
-                        errorMessage += "an unknown error occurred.";
-                    }
-                    break;
-                }
-
-                channel->addMessage(makeSystemMessage(errorMessage));
-            });
-
-        return "";
-    });
+    this->registerCommand("/marker", &commands::marker);
 
     this->registerCommand("/streamlink", [](const QStringList &words,
                                             ChannelPtr channel) {
