+++ src/controllers/commands/builtin/Misc.cpp
@@ -1,9 +1,14 @@
 #include "controllers/commands/builtin/Misc.hpp"
 
+#include "Application.hpp"
 #include "common/Channel.hpp"
+#include "controllers/accounts/AccountController.hpp"
 #include "controllers/commands/CommandContext.hpp"
 #include "messages/MessageBuilder.hpp"
+#include "providers/twitch/api/Helix.hpp"
+#include "providers/twitch/TwitchAccount.hpp"
 #include "providers/twitch/TwitchChannel.hpp"
+#include "util/FormatTime.hpp"
 #include "util/Twitch.hpp"
 
 #include <QDesktopServices>
@@ -182,6 +187,84 @@ QString clip(const CommandContext &ctx)
     }
 
     ctx.twitchChannel->createClip();
+
+    return "";
+}
+
+QString marker(const CommandContext &ctx)
+{
+    if (ctx.channel == nullptr)
+    {
+        return "";
+    }
+
+    if (ctx.twitchChannel == nullptr)
+    {
+        ctx.channel->addMessage(makeSystemMessage(
+            "The /marker command only works in Twitch channels"));
+        return "";
+    }
+
+    // Avoid Helix calls without Client ID and/or OAuth Token
+    if (getApp()->accounts->twitch.getCurrent()->isAnon())
+    {
+        ctx.channel->addMessage(makeSystemMessage(
+            "You need to be logged in to create stream markers!"));
+        return "";
+    }
+
+    // Exact same message as in webchat
+    if (!ctx.twitchChannel->isLive())
+    {
+        ctx.channel->addMessage(makeSystemMessage(
+            "You can only add stream markers during live streams. Try "
+            "again when the channel is live streaming."));
+        return "";
+    }
+
+    auto arguments = ctx.words;
+    arguments.removeFirst();
+
+    getHelix()->createStreamMarker(
+        // Limit for description is 140 characters, webchat just crops description
+        // if it's >140 characters, so we're doing the same thing
+        ctx.twitchChannel->roomId(), arguments.join(" ").left(140),
+        [channel{ctx.channel},
+         arguments](const HelixStreamMarker &streamMarker) {
+            channel->addMessage(makeSystemMessage(
+                QString("Successfully added a stream marker at %1%2")
+                    .arg(formatTime(streamMarker.positionSeconds))
+                    .arg(streamMarker.description.isEmpty()
+                             ? ""
+                             : QString(": \"%1\"")
+                                   .arg(streamMarker.description))));
+        },
+        [channel{ctx.channel}](auto error) {
+            QString errorMessage("Failed to create stream marker - ");
+
+            switch (error)
+            {
+                case HelixStreamMarkerError::UserNotAuthorized: {
+                    errorMessage +=
+                        "you don't have permission to perform that action.";
+                }
+                break;
+
+                case HelixStreamMarkerError::UserNotAuthenticated: {
+                    errorMessage += "you need to re-authenticate.";
+                }
+                break;
+
+                // This would most likely happen if the service is down, or if the JSON payload returned has changed format
+                case HelixStreamMarkerError::Unknown:
+                default: {
+                    errorMessage += "an unknown error occurred.";
+                }
+                break;
+            }
+
+            channel->addMessage(makeSystemMessage(errorMessage));
+        });
 
     return "";
 }
