+++ src/controllers/commands/builtin/Misc.hpp
@@ -17,5 +17,6 @@ QString user(const CommandContext &ctx);
 QString requests(const CommandContext &ctx);
 QString lowtrust(const CommandContext &ctx);
 QString clip(const CommandContext &ctx);
+QString marker(const CommandContext &ctx);
 
 }  // namespace chatterino::commands
